feat(test runner): new runner version initial commit
