--- scripts/test_runner/adapter_mapping.csv
@@ -1,2 +1,0 @@
-platform;vid;pid
-stm32;0x0403;0x6010
