+++ scripts/test_runner/lt_test_runner/adapter_mapping.csv
@@ -1,0 +1,3 @@
+platform;vid;pid
+stm32_f439zi;0x0403;0x6010
+stm32_l432kc;0x0403;0x6011
