+++ scripts/test_runner/lt_test_runner/ts11-jtag.cfg
@@ -1,0 +1,49 @@
+#
+# TS11 config JTAG
+# Do not specify VID and PID here. It will be specified by the test runner.
+#
+
+adapter driver ftdi
+adapter speed 2000
+transport select jtag
+
+ftdi_channel 0
+ftdi_layout_init 0x0218 0x731B
+## Signal        Data    Direction       Notes
+## TCK           0       1 (out)
+## TDI           0       1 (out)
+## TDO           0       0 (in)
+## TMS           1       1 (out)
+## SRST          1       1 (out)
+## UNUSED        0       0 (in)
+## UNUSED        0       0 (in)
+## UNUSED        0       0 (in)
+#
+## SPI_EN        0       1 (out)
+## PLTF_PWR_EN   1       1 (out)
+## PWR_FLT       0       0 (in)
+## SPI_EN_BUS    0       0 (in)
+## LED_R         0       1 (out)
+## LED_G         0       1 (out)
+## LED_B         0       1 (out)
+## Unused        0       0 (in)
+
+# We have either read-only or write-only pins.
+# Therefore, no -oe is needed. We do not want to allow changing the direction
+# on the OpenOCD CLI/telnet.
+
+# Enable SPI interface for the platform.
+ftdi layout_signal SPI_EN      -data 0x0100
+# Enable platform power. Platform power has to be enabled directly in the config, 
+# otherwise initialization may fail.
+ftdi layout_signal PLTF_PWR_EN -data 0x0200
+
+# Power fault detection.
+ftdi layout_signal PWR_FLT     -input 0x0400
+# SPI presence detection. If any platform board enables SPI access, this bus goes high.
+ftdi_layout_signal SPI_EN_BUS  -input 0x0800
+
+# Fancy disco LEDs.
+ftdi layout_signal LED_R       -data 0x1000
+ftdi layout_signal LED_G       -data 0x2000
+ftdi layout_signal LED_B       -data 0x4000
