--- scripts/test_runner/ts11-jtag.cfg
@@ -1,43 +1,0 @@
-#
-# TS11 config JTAG
-#
-
-adapter driver ftdi
-ftdi_vid_pid 0x0403 0x6010
-adapter speed 2000
-# JTAG
-transport select jtag
-
-ftdi_channel 0
-# 0018 fffb
-ftdi_layout_init 0x0F18 0x7F1B
-## Signal        Data    Direction       Notes
-## TCK           0       1 (out)
-## TDI           0       1 (in)
-## TDO           0       0 (in)
-## TMS           1       1 (out)
-## SRST          1       1 (out)
-## UNUSED        0       0 (in)
-## UNUSED        0       0 (in)
-## UNUSED        0       0 (in)
-#
-## SPI_EN        1       1 (out)
-## PLTF_PWR_EN   1       1 (out)
-## PWR_FLT       1       1 (out)
-## SPI_EN_BUS    1       1 (out)
-## PLEXILED_R    0       1 (out)
-## PLEXILED_G    0       1 (out)
-## PLEXILED_B    0       1 (out)
-## Unused        0       0 (in)
-
-ftdi_layout_signal SPI_EN      -data 0x0100 -oe 0x0100
-ftdi_layout_signal PLTF_PWR_EN -data 0x0200 -oe 0x0200
-ftdi_layout_signal PWR_FLT     -data 0x0400 -oe 0x0400
-ftdi_layout_signal SPI_EN_BUS  -data 0x0800 -oe 0x0800
-ftdi_layout_signal LED_R       -data 0x1000 -oe 0x1000
-ftdi_layout_signal LED_G       -data 0x2000 -oe 0x2000
-ftdi_layout_signal LED_B       -data 0x4000 -oe 0x4000
-
-ftdi_layout_signal CLEANUP     -data 0x0018 -oe 0x001B
-
-#ftdi_tdo_sample_edge falling
